Cartoonify_AI/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0
langchain>=0.1.0
streamlit-webrtc>=0.45.0
av>=10.0.0
onnxruntime>=1.15.0